06_dashboard/requirements_dashboard.txt
dash
plotly
boto3
pandas
python-dotenv
pyarrow
